Android recommendationCampingZones_AfterSignUp commit_CDH 2.0

+++ DYCampmate/src/main/java/com/Campmate/DYCampmate/repository/CustomerRepo.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 //JPA의 인터페이스 (DB 조회/저장 등)
 //JpaRepository를 상속하면 기본 CRUD 제공됨
-public interface CustomerRepo extends  JpaRepository<CustomerEntity,Integer> {
+public interface CustomerRepo extends  JpaRepository<CustomerEntity,Long> {
     Optional<CustomerEntity> findByCustomerId(String customerId);
     boolean existsByCustomerId(String customerId);
     Optional<CustomerEntity> findByEmail(String email);
